:recycle: api: refactor certification-attestation-repository to certificate-repository
