+++ api/lib/domain/usecases/index.js
@@ -36,7 +36,6 @@ const dependencies = {
   campaignCsvExportService: require('../../domain/services/campaign-csv-export-service'),
   certificateRepository: require('../../infrastructure/repositories/certificate-repository'),
   certificationAssessmentRepository: require('../../infrastructure/repositories/certification-assessment-repository'),
-  certificationAttestationRepository: require('../../infrastructure/repositories/certification-attestation-repository'),
   certificationAttestationPdf: require('../../infrastructure/utils/pdf/certification-attestation-pdf'),
   certificationBadgesService: require('../../domain/services/certification-badges-service'),
   certificationCandidateRepository: require('../../infrastructure/repositories/certification-candidate-repository'),
