+++ api/lib/infrastructure/repositories/certificate-repository.js
@@ -2,6 +2,7 @@ const _ = require('lodash');
 const { knex } = require('../../../db/knex-database-connection');
 const PrivateCertificate = require('../../domain/models/PrivateCertificate');
 const ShareableCertificate = require('../../domain/models/ShareableCertificate');
+const CertificationAttestation = require('../../domain/models/CertificationAttestation');
 const CertifiedBadges = require('../../../lib/domain/read-models/CertifiedBadges');
 const { NotFoundError } = require('../../../lib/domain/errors');
 const competenceTreeRepository = require('./competence-tree-repository');
@@ -64,6 +65,61 @@ module.exports = {
     const certifiedBadges = await _getCertifiedBadges(shareableCertificateDTO.id);
 
     return _toDomainForShareableCertificate({ shareableCertificateDTO, competenceTree, certifiedBadges });
+  },
+
+  async getCertificationAttestation(id) {
+    const certificationCourseDTO = await _selectCertificationAttestations()
+      .where('certification-courses.id', '=', id)
+      .groupBy('certification-courses.id', 'sessions.id', 'assessments.id', 'assessment-results.id')
+      .first();
+
+    if (!certificationCourseDTO) {
+      throw new NotFoundError(`There is no certification course with id "${id}"`);
+    }
+
+    const competenceTree = await competenceTreeRepository.get();
+    const certifiedBadges = await _getCertifiedBadges(certificationCourseDTO.id);
+
+    return _toDomainForCertificationAttestation({ certificationCourseDTO, competenceTree, certifiedBadges });
+  },
+
+  async findByDivisionForScoIsManagingStudentsOrganization({ organizationId, division }) {
+    const certificationCourseDTOs = await _selectCertificationAttestations()
+      .select({ organizationLearnerId: 'organization-learners.id' })
+      .innerJoin('certification-candidates', function () {
+        this.on({ 'certification-candidates.sessionId': 'certification-courses.sessionId' }).andOn({
+          'certification-candidates.userId': 'certification-courses.userId',
+        });
+      })
+      .innerJoin('organization-learners', 'organization-learners.id', 'certification-candidates.organizationLearnerId')
+      .innerJoin('organizations', 'organizations.id', 'organization-learners.organizationId')
+      .where({
+        'organization-learners.organizationId': organizationId,
+        'organization-learners.isDisabled': false,
+      })
+      .whereRaw('LOWER("organization-learners"."division") = ?', division.toLowerCase())
+      .whereRaw('"certification-candidates"."userId" = "certification-courses"."userId"')
+      .whereRaw('"certification-candidates"."sessionId" = "certification-courses"."sessionId"')
+      .modify(_checkOrganizationIsScoIsManagingStudents)
+      .groupBy(
+        'organization-learners.id',
+        'certification-courses.id',
+        'sessions.id',
+        'assessments.id',
+        'assessment-results.id'
+      )
+      .orderBy('certification-courses.createdAt', 'DESC');
+
+    const competenceTree = await competenceTreeRepository.get();
+
+    const mostRecentCertificationsPerOrganizationLearner =
+      _filterMostRecentCertificationCoursePerOrganizationLearner(certificationCourseDTOs);
+    return _(mostRecentCertificationsPerOrganizationLearner)
+      .orderBy(['lastName', 'firstName'], ['asc', 'asc'])
+      .map((certificationCourseDTO) => {
+        return _toDomainForCertificationAttestation({ certificationCourseDTO, competenceTree, certifiedBadges: [] });
+      })
+      .value();
   },
 };
 
@@ -181,6 +237,53 @@ function _filterMostRecentValidatedAssessmentResult(qb) {
     .where('assessment-results.status', AssessmentResult.status.VALIDATED);
 }
 
+function _selectCertificationAttestations() {
+  return knex
+    .select({
+      id: 'certification-courses.id',
+      firstName: 'certification-courses.firstName',
+      lastName: 'certification-courses.lastName',
+      birthdate: 'certification-courses.birthdate',
+      birthplace: 'certification-courses.birthplace',
+      isPublished: 'certification-courses.isPublished',
+      userId: 'certification-courses.userId',
+      date: 'certification-courses.createdAt',
+      deliveredAt: 'sessions.publishedAt',
+      verificationCode: 'certification-courses.verificationCode',
+      certificationCenter: 'sessions.certificationCenter',
+      maxReachableLevelOnCertificationDate: 'certification-courses.maxReachableLevelOnCertificationDate',
+      pixScore: 'assessment-results.pixScore',
+      assessmentResultId: 'assessment-results.id',
+      competenceMarks: knex.raw(`
+        json_agg(
+          json_build_object('score', "competence-marks".score, 'level', "competence-marks".level, 'competence_code', "competence-marks"."competence_code")
+          ORDER BY "competence-marks"."competence_code" asc
+        )`),
+    })
+    .from('certification-courses')
+    .join('assessments', 'assessments.certificationCourseId', 'certification-courses.id')
+    .join('assessment-results', 'assessment-results.assessmentId', 'assessments.id')
+    .join('competence-marks', 'competence-marks.assessmentResultId', 'assessment-results.id')
+    .join('sessions', 'sessions.id', 'certification-courses.sessionId')
+    .modify(_filterMostRecentValidatedAssessmentResult)
+    .where('certification-courses.isPublished', true)
+    .where('certification-courses.isCancelled', false);
+}
+
+function _checkOrganizationIsScoIsManagingStudents(qb) {
+  return qb.where('organizations.type', 'SCO').where('organizations.isManagingStudents', true);
+}
+
+function _filterMostRecentCertificationCoursePerOrganizationLearner(DTOs) {
+  const groupedByOrganizationLearner = _.groupBy(DTOs, 'organizationLearnerId');
+
+  const mostRecent = [];
+  for (const certificationsForOneOrganizationLearner of Object.values(groupedByOrganizationLearner)) {
+    mostRecent.push(certificationsForOneOrganizationLearner[0]);
+  }
+  return mostRecent;
+}
+
 function _toDomainForPrivateCertificate({ certificationCourseDTO, competenceTree, certifiedBadges }) {
   if (competenceTree) {
     const competenceMarks = _.compact(certificationCourseDTO.competenceMarks).map(
@@ -219,5 +322,24 @@ function _toDomainForShareableCertificate({ shareableCertificateDTO, competenceT
     ...shareableCertificateDTO,
     resultCompetenceTree,
     certifiedBadgeImages: certifiedBadges,
+  });
+}
+
+function _toDomainForCertificationAttestation({ certificationCourseDTO, competenceTree, certifiedBadges }) {
+  const competenceMarks = _.compact(certificationCourseDTO.competenceMarks).map(
+    (competenceMark) => new CompetenceMark({ ...competenceMark })
+  );
+
+  const resultCompetenceTree = ResultCompetenceTree.generateTreeFromCompetenceMarks({
+    competenceTree,
+    competenceMarks,
+    certificationId: certificationCourseDTO.id,
+    assessmentResultId: certificationCourseDTO.assessmentResultId,
+  });
+
+  return new CertificationAttestation({
+    ...certificationCourseDTO,
+    resultCompetenceTree,
+    certifiedBadges,
   });
 }
